% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ascat.R
\name{read_ascat_nc}
\alias{read_ascat_nc}
\title{Title}
\usage{
read_ascat_nc(
  file_name,
  parameter,
  rounding = c(NA, "secs", "mins", "hours"),
  bounding_date_times = NA,
  bounding_domain = NA,
  resample_domain = NA,
  lonlat_to_id = FALSE
)
}
\arguments{
\item{resample_domain}{}
}
\value{

}
\description{
Title
}
